front-end/src/app/shared/desplegable-servicio-san/servicio-san.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { ServicioSanService } from './servicio-san.service';
import { TestModule } from '../../core/mocks/test.module';
import { ApiService } from '../../core/services/restClient/api.service';
import { configureTestSuite } from '../../core/mocks/testing';
import { ServicioSan } from '../data-models/servicio-san';
import { of } from 'rxjs/internal/observable/of';

describe('ServicioSanService', () => {
  let apiService: any;
  let service: ServicioSanService;

  configureTestSuite(() => {
    apiService = jasmine.createSpyObj('ApiService', ['get']);
    TestBed.configureTestingModule({
      imports: [TestModule],
      providers: [ServicioSanService,
      {provide: ApiService, useValue: apiService}]
    });
    service = TestBed.get(ServicioSanService);
  });

  beforeEach(() => {
    apiService.get.calls.reset();
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  describe('El metodo getAll()', () => {
    it('invoca al metodo get del ApiService', () => {
      service.getAll();

      expect(apiService.get).toHaveBeenCalledWith('servicio-san');
      expect(apiService.get).toHaveBeenCalledTimes(1);
    });

    it('devuelve un observable con los datos obtenidos por el apiservice.get', () => {
      const esperado = of([{} as ServicioSan]);
      apiService.get.and.returnValue(esperado);

      const resultado = service.getAll();

      expect(resultado).toEqual(esperado);
    });
  });
});
